Android 15.0.0 Release 32 (BP1A.250505.005)


# -----BEGIN PGP SIGNATURE-----
#
# iF0EABECAB0WIQRDQNE1cO+UXoOBCWTorT+BmrEOeAUCaBqG7wAKCRDorT+BmrEO
# eIvuAJ9+WvHO4Lqn+2pvGsw9vgZEnqr+XwCeJqVT+9TpFAqBLShgGwY0pygbUfs=
# =B498
# -----END PGP SIGNATURE-----
# gpg: Signature faite le mar 06 mai 2025 18:02:23 EDT
# gpg:                avec la clef DSA 4340D13570EF945E83810964E8AD3F819AB10E78
# gpg: Impossible de vérifier la signature : Pas de clef publique

+++ payload_consumer/filesystem_verifier_action.h
@@ -176,7 +176,7 @@ class FilesystemVerifierAction : public InstallPlanAction {
   // Cumulative sum of partition sizes. Used for progress report.
   // This vector will always start with 0, and end with total size of all
   // partitions.
-  std::vector<size_t> partition_weight_;
+  std::vector<uint64_t> partition_weight_;
 
   DISALLOW_COPY_AND_ASSIGN(FilesystemVerifierAction);
 };
